Make all MP3 buffer errors immediately stop()
I'll try to debug and fix this later, but for now, this is a work around
for some MP3s I'm using on a project that fail their final decode at the
end of the file after playing successfully. Then the loop() method never
detects that the MP3 is done.